refactor(add): extract finalization.py + split add_files into explicit phases (#624)
Extract batch finalization (STAC-write + backend coordination) out of add.py
into a new portolan_cli/finalization.py, and split add_files into real,
independently testable phase functions.

finalization.py owns finalize_items, _finalize_with_backend,
_recompute_collection_extent_with_multibbox and their helpers
(_get_or_create_collection, _add_prepared_items_to_collection,
_collect_parquet_metadata_from_disk, _warn_about_stale_assets,
_ensure_partition_metadata, _save_collection_with_links (+link helpers),
_batch_update_versions, _asset_freshness_fields). The dependency edge is
one-directional (add -> finalization); catalog/backends/version_ops/bbox
imports stay function-local so add -> finalization -> catalog -> add cannot
re-form a load-time cycle. add.py re-exports the moved names so existing
`from portolan_cli.add import ...` and `patch("portolan_cli.add.finalize_items")`
sites keep working; external.py imports _save_collection_with_links from the
new module.

add_files is now a thin orchestrator over explicit phases: _collect_files_for_add
(1) -> _phase_process (2, with module-level _prepare_single_file /
_prepare_multilayer_file + _PrepareOptions/_ProcessResult, replacing the deep
prepare_single_file closure) -> finalize_items (2.5) ->
_process_deferred_non_geo_files (3) -> _recompute_stats_for_collections (3.5).

Complexity: add_files C(19)->A(4), finalize_items C(20)->A(5),
_recompute_collection_extent_with_multibbox C(20)->A(5); max add-path nesting
depth drops well below 9. Behavior preserved; multilayer error semantics kept
identical (broad outer try). Package xenon/ruff/mypy/import-linter green.

Part of #618.
